--- app/src/main/java/com/example/generationapp/ScreenSaverActivity.java
@@ -1,90 +1,0 @@
-package com.example.generationapp;
-
-import android.annotation.SuppressLint;
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.media.MediaPlayer;
-import android.net.Uri;
-import android.os.Environment;
-import android.provider.MediaStore;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.MotionEvent;
-import android.view.View;
-import android.widget.MediaController;
-import android.widget.Toast;
-import android.widget.VideoView;
-
-import java.io.File;
-
-public class ScreenSaverActivity extends AppCompatActivity {
-
-    private static final String TAG = "";
-    VideoView videoView;
-
-
-    @SuppressLint("ClickableViewAccessibility")
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_screen_saver);
-
-//       File file  = new File(Environment.getExternalStorageDirectory()+"/Generation Videos");
-//        Log.e(TAG, "Folder Name " + file.getName());
-
-        VideoStreamer();
-
-        videoView.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View view, MotionEvent motionEvent) {
-
-                startActivity(new Intent(ScreenSaverActivity.this, WebActivity.class));
-                finish();
-                return false;
-            }
-        });
-    }
-
-    private void VideoStreamer() {
-        videoView = findViewById(R.id.video_view);
-        String videoPath = "android.resource://" + getPackageName() + "/" + R.raw.video;
-        Uri uri = Uri.parse(videoPath);
-        videoView.setVideoURI(uri);
-
-        videoView.setOnCompletionListener(new MediaPlayer.OnCompletionListener() {
-            @Override
-            public void onCompletion(MediaPlayer mediaPlayer) {
-
-            }
-        });
-        videoView.requestFocus();
-
-        videoView.setOnPreparedListener(new MediaPlayer.OnPreparedListener() {
-            @Override
-            public void onPrepared(MediaPlayer mediaPlayer) {
-                mediaPlayer.setLooping(true);
-                videoView.start();
-
-            }
-        });
-    }
-
-    @Override
-    public void onBackPressed() {
-        new AlertDialog.Builder(this)
-                .setIcon(R.drawable.ic_warning_icon)
-                .setTitle("Exit App")
-                .setMessage("Are you sure you want to close this App?")
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        finish();
-                    }
-
-                })
-                .setNegativeButton("No", null)
-                .show();
-    }
-}
