+++ app/src/main/res/layout/activity_screen_saver.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScreenSaverActivity">
+    tools:context=".Controller.ScreenSaverActivity">
 
     <VideoView
         android:id="@+id/video_view"
