+++ app/src/main/res/layout/activity_slider.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SliderActivity"
+    tools:context=".Controller.SliderActivity"
     android:background="@color/colorPrimary"
     android:id="@+id/relative_layout">
