+++ app/src/main/res/layout/activity_splash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashActivity">
+    tools:context=".Controller.SplashActivity">
 
     <ImageView
         android:layout_width="match_parent"
@@ -25,7 +25,7 @@
         android:layout_below="@id/logo"
         android:layout_centerHorizontal="true"
         android:layout_centerInParent="true"
-        android:visibility="visible"
+        android:visibility="invisible"
         android:theme="@style/whiteAccent"/>
