+++ app/src/main/res/layout/activity_web.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WebActivity"
+    tools:context=".Controller.WebActivity"
     android:id="@+id/relative_layout">
 
     <android.support.v4.widget.SwipeRefreshLayout
